backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
python-dotenv==1.0.0
pymongo==4.6.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
protobuf==4.25.1
sentence-transformers==2.2.2
sacremoses==0.0.53
functions-framework==3.*
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
firebase-functions==0.1.0
mangum==0.17.0